+++ torchmetrics/text/bleu.py
@@ -59,6 +59,8 @@ class BLEUScore(Metric):
         and Skip-Bigram Statistics by Chin-Yew Lin and Franz Josef Och `Machine Translation Evolution`_
     """
 
+    is_differentiable = False
+    higher_is_better = True
     trans_len: Tensor
     ref_len: Tensor
     numerator: Tensor
@@ -116,7 +118,3 @@ class BLEUScore(Metric):
         return _bleu_score_compute(
             self.trans_len, self.ref_len, self.numerator, self.denominator, self.n_gram, self.smooth
         )
-
-    @property
-    def is_differentiable(self) -> bool:
-        return False